Slider value snapping
Most sliders now snap to certain values. The formula for the peak sliders is improved to be close to linear in the 0-1 range, but go to +- infinity more quickly outside that range.
Refactored wildness and worth calculations to be more unified; the same method does both by writing values to whichever object (wildParams or worthParams) is passed to it.