Fates can prevent "draw no more" effect
The Fates card's description allows the character to avoid any event, so they should be able to negate an effect that prevents further draws.

Also, this special property of potentially negating a draw-ending effect already existed in the Lottery card, but the property would be lost if the deck was exported; this has been fixed.